+++ src/pages/Home/Home.scss
@@ -25,14 +25,12 @@
 }
 
 .parrentBanner{
-        
-        display:flex;
-        justify-content:center;
-        align-items:center;
-        width:100%;
-        height:38vh;
-        margin-top:120px;
-        margin-bottom:100px;
+    display:flex;
+    justify-content:center;
+    align-items:center;
+    width:100%;
+    height:38vh;
+    transform: translateY(50%);
 
     .banner{
         background: url("../../assets/images/ready-to.png");
@@ -63,6 +61,14 @@
         height:107px;
         margin-top:170px;
     }
+
+    .canvas-background {
+        top: -220px;
+
+        canvas {
+            height: 90%;
+        }
+    }
 }
 @media screen and (max-width:600px){
     .parrentBanner{
@@ -79,8 +85,8 @@
 
 .About{
     align-items: start;
-    
     padding-top: 100px;
+
     .AboutFlex{
         justify-content: space-between;
     }
